Move mbtiles code to a separate class to allow easier removal in the future.
