+++ IsraelHiking.Web/src/application/application.module.ts
@@ -96,6 +96,7 @@ import { GlobalErrorHandler } from "./services/global-error.handler";
 import { OverpassTurboService } from "./services/overpass-turbo.service";
 import { ImageAttributionService } from "./services/image-attribution.service";
 import { PmTilesService } from "./services/pmtiles.service";
+import { MBTilesService } from "./services/mbtiles.service";
 // interactions
 import { RouteEditPoiInteraction } from "./components/intercations/route-edit-poi.interaction";
 import { RouteEditRouteInteraction } from "./components/intercations/route-edit-route.interaction";
@@ -308,6 +309,7 @@ const initializeApplication = (injector: Injector) => async () => {
             OverpassTurboService,
             ImageAttributionService,
             PmTilesService,
+            MBTilesService,
             AudioPlayerFactory,
             FileSystemWrapper,
             // eslint-disable-next-line
