+++ IsraelHiking.Web/src/application/components/dialogs/files-shares-dialog.component.ts
@@ -57,19 +57,10 @@ export class FilesSharesDialogComponent extends BaseMapComponent {
             }
             return;
         }
-        if (file.name.endsWith(".mbtiles") && offlineState.isOfflineAvailable) {
-            this.toastService.info(this.resources.openingAFilePleaseWait);
-            const dbFileName = file.name.replace(".mbtiles", ".db");
-            await this.fileService.storeFileToCache(dbFileName, file);
-            await this.databaseService.moveDownloadedDatabaseFile(dbFileName);
-            this.toastService.confirm({ type: "Ok", message: this.resources.finishedOpeningTheFile });
-            this.store.dispatch(new SetOfflineMapsLastModifiedDateAction(new Date(file.lastModified)));
-            return;
-        }
         if (file.name.endsWith(".pmtiles")) {
             this.toastService.info(this.resources.openingAFilePleaseWait);
             await this.fileService.storeFileToCache(file.name, file);
-            await this.fileService.moveDownloadedDatabaseFile(file.name);
+            await this.fileService.moveFileFromCacheToDataDirectory(file.name);
             this.toastService.confirm({ type: "Ok", message: this.resources.finishedOpeningTheFile });
             this.store.dispatch(new SetOfflineMapsLastModifiedDateAction(new Date(file.lastModified)));
             return;
