+++ IsraelHiking.Web/src/application/services/database.service.ts
@@ -1,8 +1,6 @@
 import { Injectable } from "@angular/core";
 import { Store } from "@ngxs/store";
 import { debounceTime } from "rxjs/operators";
-import { CapacitorSQLite, SQLiteDBConnection, SQLiteConnection} from "@capacitor-community/sqlite";
-import { gunzipSync } from "fflate";
 import Dexie from "dexie";
 import deepmerge from "deepmerge";
 import maplibregl from "maplibre-gl";
@@ -10,6 +8,7 @@ import maplibregl from "maplibre-gl";
 import { LoggingService } from "./logging.service";
 import { RunningContextService } from "./running-context.service";
 import { PmTilesService } from "./pmtiles.service";
+import { MBTilesService } from "./mbtiles.service";
 import { POPULARITY_HEATMAP, initialState } from "../reducers/initial-state";
 import { ClearHistoryAction } from "../reducers/routes.reducer";
 import { SetSelectedPoiAction, SetSidebarAction } from "../reducers/poi.reducer";
@@ -42,26 +41,22 @@ export class DatabaseService {
     private imagesDatabase: Dexie;
     private shareUrlsDatabase: Dexie;
     private tracesDatabase: Dexie;
-    private sourceDatabases: Map<string, Promise<SQLiteDBConnection>>;
     private updating: boolean;
-    private sqlite: SQLiteConnection;
 
     constructor(private readonly loggingService: LoggingService,
                 private readonly runningContext: RunningContextService,
                 private readonly pmTilesService: PmTilesService,
+                private readonly mbtilesService: MBTilesService,
                 private readonly store: Store) {
         this.updating = false;
-        this.sourceDatabases = new Map<string, Promise<SQLiteDBConnection>>();
     }
 
     public async initialize() {
+        this.mbtilesService.initialize();
         this.stateDatabase = new Dexie(DatabaseService.STATE_DB_NAME);
         this.stateDatabase.version(1).stores({
             state: "id"
         });
-        if (this.runningContext.isCapacitor) {
-            this.sqlite = new SQLiteConnection(CapacitorSQLite);
-        }
         this.poisDatabase = new Dexie(DatabaseService.POIS_DB_NAME);
         this.poisDatabase.version(1).stores({
             pois: DatabaseService.POIS_ID_COLUMN + "," + DatabaseService.POIS_LOCATION_COLUMN,
@@ -127,27 +122,7 @@ export class DatabaseService {
         this.store.dispatch(new SetSidebarAction(false));
         const finalState = this.store.snapshot() as ApplicationState;
         await this.updateState(finalState);
-        for (const dbKey of this.sourceDatabases.keys()) {
-            await this.closeDatabase(dbKey);
-        }
-    }
-
-    public async closeDatabase(dbKey: string) {
-        this.loggingService.info("[Database] Closing " + dbKey);
-        if (!this.sourceDatabases.has(dbKey)) {
-            this.loggingService.info(`[Database] ${dbKey} was never opened`);
-            return;
-        }
-        try {
-            const db = await this.sourceDatabases.get(dbKey);
-            await db.close();
-            this.loggingService.info("[Database] Closed succefully: " + dbKey);
-            await this.sqlite.closeConnection(dbKey + ".db", true);
-            this.loggingService.info("[Database] Connection closed succefully: " + dbKey);
-            this.sourceDatabases.delete(dbKey);
-        } catch (ex) {
-            this.loggingService.error(`[Database] Unable to close ${dbKey}, ${(ex as Error).message}`);
-        }
+        await this.mbtilesService.uninitialize();
     }
 
     private async updateState(state: ApplicationState) {
@@ -167,77 +142,13 @@ export class DatabaseService {
         }
     }
 
-    private getSourceNameFromUrl(url: string) {
-        return url.replace("custom://", "").split("/")[0];
-    }
-
     public async getTile(url: string): Promise<ArrayBuffer> {
         try {
             return await this.pmTilesService.getTile(url);
         } catch (ex) {
             this.loggingService.error(`[Database] Failed to get tile from pmtiles: ${(ex as Error).message}`);
         }
-        const splitUrl = url.split("/");
-        const dbName = this.getSourceNameFromUrl(url);
-        const z = +splitUrl[splitUrl.length - 3];
-        const x = +splitUrl[splitUrl.length - 2];
-        const y = +(splitUrl[splitUrl.length - 1].split(".")[0]);
-
-        return this.getTileFromDatabase(dbName, z, x, y);
-    }
-
-    private async getTileFromDatabase(dbName: string, z: number, x: number, y: number): Promise<ArrayBuffer> {
-        const db = await this.getDatabase(dbName);
-
-        const params = [z, x, Math.pow(2, z) - y - 1];
-        const queryresults = await db.query("SELECT HEX(tile_data) as tile_data_hex FROM tiles " +
-                "WHERE zoom_level = ? AND tile_column = ? AND tile_row = ? limit 1",
-                params);
-        if (queryresults.values.length !== 1) {
-            throw new Error("Unable to get tile from database");
-        }
-        const hexData = queryresults.values[0].tile_data_hex;
-        let binData = new Uint8Array(hexData.match(/.{1,2}/g).map((byte: string) => parseInt(byte, 16)));
-        const isGzipped = binData[0] === 0x1f && binData[1] === 0x8b;
-        if (isGzipped) {
-            binData = gunzipSync(binData);
-        }
-        return binData.buffer;
-    }
-
-    private async getDatabase(dbName: string): Promise<SQLiteDBConnection> {
-        if (this.sourceDatabases.has(dbName)) {
-            try {
-                const db = await this.sourceDatabases.get(dbName);
-                return db;
-            } catch (ex) {
-                this.loggingService.error(`[Database] There's a problem with the connection to ${dbName}, ${(ex as Error).message}`);
-            }
-        }
-        this.loggingService.info(`[Database] Creating connection to ${dbName}`);
-        this.sourceDatabases.set(dbName, this.createConnection(dbName));
-        return this.sourceDatabases.get(dbName);
-    }
-
-    private async createConnection(dbName: string) {
-        try {
-            const dbPromise = this.sqlite.createConnection(dbName + ".db", false, "no-encryption", 1, true);
-            const db = await dbPromise;
-            this.loggingService.info(`[Database] Connection created succefully to ${dbName}`);
-            await db.open();
-            this.loggingService.info(`[Database] Connection opened succefully: ${dbName}`);
-            return db;
-        } catch (ex) {
-            this.loggingService.error(`[Database] Failed opening ${dbName}, ${(ex as Error).message}`);
-            throw ex;
-        }
-    }
-
-    public async moveDownloadedDatabaseFile(dbFileName: string) {
-        await this.closeDatabase(dbFileName.replace(".db", ""));
-        this.loggingService.info(`[Database] Starting moving file ${dbFileName}`);
-        await this.sqlite.moveDatabasesAndAddSuffix("cache", [dbFileName]);
-        this.loggingService.info(`[Database] Finished moving file ${dbFileName}`);
+        return this.mbtilesService.getTile(url);
     }
 
     public storePois(pois: GeoJSON.Feature[]): Promise<any> {
@@ -341,12 +252,5 @@ export class DatabaseService {
             storedState.gpsState = initialState.gpsState;
         }
         return storedState;
-    }
-
-    public async migrateDatabasesIfNeeded(): Promise<void> {
-        this.loggingService.info("[Database] Starting migrating old databases using sqlite plugin");
-        await this.sqlite.moveDatabasesAndAddSuffix("default", ["Contour.db", "IHM.db", "TerrainRGB.db"]);
-        const databases = await this.sqlite.getDatabaseList();
-        this.loggingService.info("[Database] Finished migrating old databases using sqlite plugin, " + JSON.stringify(databases.values));
     }
 }
