+++ IsraelHiking.Web/src/application/services/file.service.ts
@@ -354,31 +354,7 @@ export class FileService {
         this.loggingService.info(`[Files] Finished downloading and writing file to cache, file name ${fileName}`);
     }
 
-    public async renameOldDatabases(): Promise<boolean> {
-        if (!this.runningContextService.isCapacitor) {
-            return false;
-        }
-        let filesExist = false;
-        const filePrefix = this.runningContextService.isIos ? "" : "databases/";
-        const originFolder = this.runningContextService.isIos
-            ? this.fileSystemWrapper.documentsDirectory
-            : this.fileSystemWrapper.applicationStorageDirectory;
-        for (const fileName of ["Contour.mbtiles", "IHM.mbtiles", "TerrainRGB.mbtiles"]) {
-            const fullFileName = filePrefix + fileName;
-            this.loggingService.info(`[Files] Checking if database file exists: ${fullFileName}`);
-            try {
-                const fileExists = await this.fileSystemWrapper.checkFile(originFolder, fullFileName);
-                if (!fileExists) { continue; }
-                this.loggingService.info(`[Files] Statring renaming database: ${fullFileName}`);
-                await this.fileSystemWrapper.moveFile(originFolder, fullFileName, originFolder, fullFileName.replace(".mbtiles", ".db"));
-                this.loggingService.info(`[Files] Finished renaming database: ${fullFileName}`);
-                filesExist = true;
-            } catch { } // eslint-disable-line
-        }
-        return filesExist;
-    }
-
-    public async moveDownloadedDatabaseFile(fileName: string): Promise<void> {
+    public async moveFileFromCacheToDataDirectory(fileName: string): Promise<void> {
         await this.fileSystemWrapper.moveFile(this.fileSystemWrapper.cacheDirectory, fileName, this.fileSystemWrapper.dataDirectory, fileName);
     }
 }
